Integrate upcoming courses and improve registration
