+++ src/components/Navbar.tsx
@@ -9,8 +9,7 @@ const Navbar = () => {
     { name: "About", href: "#about" },
     { name: "Solutions", href: "#solutions" },
     { name: "Consulting", href: "#consulting" },
-    { name: "Courses", href: "#courses" },
-    { name: "Upcoming Courses", href: "/upcoming-courses", isPage: true },
+    { name: "Upcoming Courses", href: "#upcoming-courses" },
     { name: "Contact", href: "#contact" },
   ];
